fix: cast to returned version value to int
- the new methods return a long instead of an int as the old ones did

+++ gestalt-module/src/main/java/org/terasology/gestalt/naming/Version.java
@@ -87,15 +87,15 @@ public final class Version implements Comparable<Version> {
     }
 
     public int getMajor() {
-        return semver.majorVersion();
+        return (int) semver.majorVersion();
     }
 
     public int getMinor() {
-        return semver.minorVersion();
+        return (int) semver.minorVersion();
     }
 
     public int getPatch() {
-        return semver.patchVersion();
+        return (int) semver.patchVersion();
     }
 
     /**
